pyproject.toml
[project]
name = "nishchinto-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Fardeen Ehsan",email = "fardeen.es7@gmail.com"}
]
requires-python = "^3.12"
dependencies = [
    "django (==6.0.4)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "celery (>=5.6.3,<6.0.0)",
    "redis (>=7.4.0,<8.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-pg-zero-downtime-migrations (>=0.19,<0.20)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-environ (>=0.13.0,<0.14.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "sentry-sdk (>=2.58.0,<3.0.0)",
    "django-allauth (>=65.16.1,<66.0.0)",
    "dj-rest-auth (>=7.2.0,<8.0.0)",
    "opentelemetry-api (>=1.41.0,<2.0.0)",
    "opentelemetry-sdk (>=1.41.0,<2.0.0)",
    "opentelemetry-instrumentation-django (>=0.62b0,<0.63)",
    "requests (>=2.33.1,<3.0.0)",
    "cryptography (>=46.0.7,<47.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "django-storages[s3] (>=1.14.0,<2.0.0)",
    "boto3 (>=1.38.0,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "django-filter (>=25.1,<26.0)",
    "django-celery-beat (>=2.7.0,<3.0.0)",
    "meilisearch (>=0.31.0,<1.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
